Laravel 12 compatibility (#26)
Hi there,

I saw issue #25 was still open, I tried to add Laravel 12 compatibility.

I tried to generate a full set of github actions for it, I'm not sure if
that's correct as it was my first usage of Actions feature.
Not sure what to do with Larastan Type Errors and BC job.

Please do not hesitate to ask for changes or Edit yourself.

Cheers

---------

Co-authored-by: Michael Rubél <contact@observer.name>
Co-authored-by: Michael Rubél <michael@laravel.software>